fix: skip memory-intensive tests
