+++ test/integration_tests/test_reuse.py
@@ -15,8 +15,8 @@
 import os
 import unittest
 from pathlib import Path
-from test.markers import windows_skip_marker
-from test.util import get_logical_query_plan, load_udfs_for_testing, shutdown_ray, memory_skip_marker
+from test.markers import memory_skip_marker, windows_skip_marker
+from test.util import get_logical_query_plan, load_udfs_for_testing, shutdown_ray
 
 from eva.catalog.catalog_manager import CatalogManager
 from eva.configuration.constants import EVA_ROOT_DIR
